+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - `--max-workers` run mode flag to control global maximum of concurrent downloads (was previously set by `GLOBAL_MAX_WORKERS = 12`). Updated documentation and tests accordingly
 
 ### Changed
+- Consolidated output folder creation utilities between parsing and generating, and moved parser logs to standard output/[screenplay]/logs location
 - Changing behavior of `--tts-provider` for `sts-tts-provider-yaml` flag to explicitly populate TTS provider fields to remove the necessity of `--tts-provider` flag for `sts-generate-audio` 
 - Naming more consistent and descriptive in commands, code, and documentation
   - "provider" -> "TTS provider"